recruiting/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
gunicorn==21.2.0
textblob==0.17.1
python-dotenv==1.0.0
facebook-business==19.0.0
requests==2.31.0
itsdangerous==2.2.0
